Move bazel-diff download code to diff_utils.py
Consolidate all bazel-diff machinery (download, hash generation, target
computation) in diff_utils.py. check_release_lib.py now imports
download_bazel_diff from diff_utils instead of defining it locally.
